# docs/openapi/crm-v1-openapi.yaml
openapi: 3.0.3
info:
  title: Canicula CRM V1 API
  version: 0.1.0
  description: Project-based key account CRM V1 interface contract.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /api/receivable-plans:
    get:
      tags: [Receivables]
      summary: List receivable plans
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: account_id
          in: query
          schema:
            type: integer
            format: int64
        - name: opportunity_id
          in: query
          schema:
            type: integer
            format: int64
        - name: contract_id
          in: query
          schema:
            type: integer
            format: int64
        - name: receivable_status
          in: query
          schema:
            type: string
        - name: plan_stage
          in: query
          schema:
            type: string
        - name: owner_user_id
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Receivable plan list.
    post:
      tags: [Receivables]
      summary: Create receivable plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Receivable plan detail.
  /api/receivable-plans/{planId}:
    get:
      tags: [Receivables]
      summary: Get receivable plan detail
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Receivable plan detail.
    patch:
      tags: [Receivables]
      summary: Update receivable plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Receivable plan detail.
  /api/receivable-plans/{planId}/terminate:
    post:
      tags: [Receivables]
      summary: Terminate receivable plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Receivable plan detail.
  /api/receivable-plans/{planId}/payments:
    get:
      tags: [Receivables]
      summary: List payments under a receivable plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment list under the receivable plan.
  /api/receivable-plans/{planId}/follow-ups:
    get:
      tags: [Receivables]
      summary: List receivable follow-ups
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Receivable follow-up list.
    post:
      tags: [Receivables]
      summary: Create receivable follow-up
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Receivable follow-up detail.
...
  /api/reconciliations/workbench:
    get:
      tags: [Reconciliations]
      summary: Load invoice-payment reconciliation workbench
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: account_id
          in: query
          schema:
            type: integer
            format: int64
        - name: opportunity_id
          in: query
          schema:
            type: integer
            format: int64
        - name: contract_id
          in: query
          schema:
            type: integer
            format: int64
        - name: pending_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Workbench summary, pending invoices, pending payments, and recent reconciliations.
  /api/reconciliations:
    get:
      tags: [Reconciliations]
      summary: List reconciliation records
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: account_id
          in: query
          schema:
            type: integer
            format: int64
        - name: opportunity_id
          in: query
          schema:
            type: integer
            format: int64
        - name: contract_id
          in: query
          schema:
            type: integer
            format: int64
        - name: invoice_id
          in: query
          schema:
            type: integer
            format: int64
        - name: payment_id
          in: query
          schema:
            type: integer
            format: int64
        - name: reconciliation_status
          in: query
          schema:
            type: string
        - name: active_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Reconciliation record list.
    post:
      tags: [Reconciliations]
      summary: Create reconciliation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invoice_id, payment_id, reconciled_amount]
              properties:
                invoice_id:
                  type: integer
                  format: int64
                payment_id:
                  type: integer
                  format: int64
                reconciled_amount:
                  type: number
                reconciled_at:
                  type: string
                  format: date-time
                reconcile_note:
                  type: string
      responses:
        '200':
          description: Reconciliation detail.
  /api/reconciliations/{reconciliationId}:
    get:
      tags: [Reconciliations]
      summary: Get reconciliation detail
      parameters:
        - name: reconciliationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reconciliation detail.
  /api/reconciliations/{reconciliationId}/void:
    post:
      tags: [Reconciliations]
      summary: Void reconciliation
      parameters:
        - name: reconciliationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [void_reason]
              properties:
                void_reason:
                  type: string
      responses:
        '200':
          description: Reconciliation detail.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...